Preprocessing modified and included to fit/predict
